requirements.txt
pandas>=2.1.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
scikit-learn>=1.3.0
mlflow>=2.7.0
apache-airflow>=2.7.0
dvc>=3.20.0
